+++ test/tests/services/test.js
@@ -1,5 +1,4 @@
 "use strict";
-require('./coinmarketcap/test');
 require('./marketCap/test');
 require('./fxConverter/test');
 require('./pushover/test');